+++ package.json
@@ -1,14 +1,29 @@
 {
-  "name": "coffeemug",
-  "module": "index.js",
-  "type": "module",
-  "devDependencies": {
-    "@types/bun": "latest"
-  },
-  "peerDependencies": {
-    "typescript": "^5.0.0"
-  },
-  "dependencies": {
-    "express": "^4.21.2"
-  }
+    "name": "coffeemug",
+    "module": "index.js",
+    "type": "module",
+    "scripts": {
+        "migrate": "drizzle-kit generate && drizzle-kit migrate",
+        "start": "tsx index.ts"
+    },
+    "devDependencies": {
+        "@types/bun": "latest",
+        "@types/express": "^5.0.0",
+        "@types/sqlite3": "^3.1.11",
+        "drizzle-kit": "^0.30.1",
+        "tsx": "^4.19.2"
+    },
+    "peerDependencies": {
+        "typescript": "^5.0.0"
+    },
+    "dependencies": {
+        "@libsql/client": "^0.14.0",
+        "dotenv": "^16.4.7",
+        "drizzle": "^1.4.0",
+        "drizzle-orm": "^0.38.3",
+        "drizzle-zod": "^0.6.1",
+        "express": "^4.21.2",
+        "sqlite3": "^5.1.7",
+        "zod": "^3.24.1"
+    }
 }
