+++ src/database/driver.ts
@@ -1,4 +1,4 @@
 import { drizzle } from "drizzle-orm/libsql";
-import { env } from "../env.js";
+import { env } from "../env";
 
 export const db = drizzle(env.dbFileName);
