+++ src/env.ts
@@ -1,7 +1,7 @@
 import "dotenv/config";
+import { EnvironmentVariables } from "./types/env";
 
-/** @type {EnvironmentVariables} */
 export const env = {
     port: Number.parseInt(process.env.PORT || "3000"),
     dbFileName: process.env.DB_FILE_NAME || ":memory:",
-};
+} satisfies EnvironmentVariables;
