+++ src/types/env.ts
@@ -1,0 +1,4 @@
+export type EnvironmentVariables = {
+    port: number;
+    dbFileName: string;
+};
